Removed some compiler warnings : - Changed all host only based C strings that were "BYTE *" to "char *" - Do not assume a "long long" is 64 bit or the only 64 bit long integer representation - read_socket/write_socket now match recv() to use a "void *" as a buffer - pre-initialize some local variables some compilers claims may be used initialized (when the code logic dictates it is untrue) - comment out some unused variables - define some variables used under some preprocessor conditions only when those conditions are met - avoid using type punned pointers (breaks GCC badly sometimes)
git-svn-id: file:///home/jj/hercules.svn/trunk@3997 956126f8-22a0-4046-8f4a-272fa8102e63

+++ tapedev.c
@@ -701,7 +701,7 @@ U16             prvblkl;                /* Length of previous block  */
     /* ISW: Determine if we are passed maxsize */
     if(dev->tdparms.maxsize>0)
     {
-        if(dev->nxtblkpos+blklen+sizeof(awshdr) > dev->tdparms.maxsize)
+        if((off_t)(dev->nxtblkpos+blklen+sizeof(awshdr)) > dev->tdparms.maxsize)
         {
             build_senseX(TAPE_BSENSE_ENDOFTAPE,dev,unitstat,code);
             return -1;
@@ -822,7 +822,7 @@ U16             prvblkl;                /* Length of previous block  */
     /* ISW: Determine if we are passed maxsize */
     if(dev->tdparms.maxsize>0)
     {
-        if(dev->nxtblkpos+sizeof(awshdr) > dev->tdparms.maxsize)
+        if((off_t)(dev->nxtblkpos+sizeof(awshdr)) > dev->tdparms.maxsize)
         {
             build_senseX(TAPE_BSENSE_ENDOFTAPE,dev,unitstat,code);
             return -1;
